refactor: Relocate sprint_row() to conftest This reduces duplication across tests for creating sample SprintBoard data
