+++ analytics/tests/datasets/test_sprint_board.py
@@ -5,56 +5,15 @@ import pytest
 from analytics.datasets.sprint_board import SprintBoard
 
 from tests.conftest import (
+    DAY_1,
+    DAY_2,
+    DAY_4,
+    DAY_5,
     json_issue_row,
     json_sprint_row,
+    sprint_row,
     write_test_data_to_file,
 )
-
-DAY_1 = "2023-11-01"
-DAY_2 = "2023-11-02"
-DAY_3 = "2023-11-03"
-DAY_4 = "2023-11-04"
-DAY_5 = "2023-11-05"
-
-
-def sprint_board_row(
-    issue: int,
-    created: str = DAY_1,
-    closed: str | None = None,
-    status: str = "In Progress",
-    points: int = 1,
-    sprint: int = 1,
-    sprint_start: str = DAY_1,
-    sprint_length: int = 2,
-) -> dict:
-    """Create a sample row of the SprintBoard dataset."""
-    # create timestamp and time delta fields
-    sprint_start_ts = pd.Timestamp(sprint_start)
-    sprint_duration = pd.Timedelta(days=sprint_length)
-    sprint_end_ts = sprint_start_ts + sprint_duration
-    created_date = pd.Timestamp(created, tz="UTC")
-    closed_date = pd.Timestamp(closed, tz="UTC") if closed else None
-    # return the sample record
-    return {
-        "issue_number": issue,
-        "issue_title": f"Issue {issue}",
-        "type": "issue",
-        "issue_body": f"Description of issue {issue}",
-        "status": "Done" if closed else status,
-        "assignees": "mickeymouse",
-        "labels": [],
-        "url": f"https://github.com/HHS/simpler-grants-gov/issues/{issue}",
-        "points": points,
-        "milestone": "Milestone 1",
-        "milestone_due_date": sprint_end_ts,
-        "milestone_description": "Milestone 1 description",
-        "sprint": f"Sprint {sprint}",
-        "sprint_start_date": sprint_start_ts,
-        "sprint_end_date": sprint_end_ts,
-        "sprint_duration": sprint_duration,
-        "created_date": created_date,
-        "closed_date": closed_date,
-    }
 
 
 class TestSprintBoard:
@@ -159,9 +118,9 @@ class TestGetSprintNameFromDate:
         """Test that correct sprint is returned if date exists in a sprint."""
         # setup - create sample dataset
         board_data = [
-            sprint_board_row(issue=1, sprint=1, sprint_start=DAY_1),
-            sprint_board_row(issue=2, sprint=1, sprint_start=DAY_1),
-            sprint_board_row(issue=3, sprint=2, sprint_start=DAY_4),
+            sprint_row(issue=1, sprint=1, sprint_start=DAY_1),
+            sprint_row(issue=2, sprint=1, sprint_start=DAY_1),
+            sprint_row(issue=3, sprint=2, sprint_start=DAY_4),
         ]
         board = SprintBoard.from_dict(board_data)
         # validation
@@ -173,8 +132,8 @@ class TestGetSprintNameFromDate:
         """The method should return None if no sprint contains the date."""
         # setup - create sample dataset
         board_data = [
-            sprint_board_row(issue=1, sprint=1, sprint_start=DAY_1),
-            sprint_board_row(issue=2, sprint=2, sprint_start=DAY_4),
+            sprint_row(issue=1, sprint=1, sprint_start=DAY_1),
+            sprint_row(issue=2, sprint=2, sprint_start=DAY_4),
         ]
         board = SprintBoard.from_dict(board_data)
         # validation
